requirements.txt
# Deep Learning / ML
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
sentence-transformers==3.0.1

# Data Science & Computer Vision
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
scipy==1.13.1
matplotlib==3.9.0
opencv-python==4.10.0.84
Pillow==10.3.0

# Web Framework
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
PyMySQL==1.1.1

# Utilities
PyMuPDF==1.24.5
textdistance==4.6.3
nltk==3.8.1
joblib==1.4.2
tqdm==4.66.4
